Refactored from yesterday.
The download interface in the teacer's Google Drive was different today.
Yesterday I didn't have to clicl the image of the file to be downloades.
Just hovering the mouse a download arrow appeared over the image.
Not today...
I had to click once. Then a 3-dot menu appeared in the right corner of the screen, with the download option.
Instead of blaming, I did it all again. Fuck!